Fix: rutas para el delete de comunidad

+++ api/controllers/communityControllers/communityController.js
@@ -5,13 +5,20 @@ import Community from "../../models/community/Community.js";
 const deleteCommunityById = async (req, res) => {
   try {
     const communityId = req.params.id;
-    const community = await Community.findByIdAndDelete(communityId);
+    const community = await Community.findById(communityId);
 
     if (!community) {
       return res.status(404).json({ message: 'Comunidad no encontrada' });
     }
 
-    res.status(200).json({ message: 'Comunidad eliminada exitosamente' });
+    if (!community.authorId.toString() !== req.user.id) {
+      return res.status(403).json({
+        message: 'No tienes permiso para eliminar esta comunidad'
+      });      
+    }
+
+    await Community.findByIdAndDelete(communityId);
+      res.status(200).json({ message: 'Comunidad eliminada exitosamente' });
   } catch (error) {
     res.status(500).json({ message: 'Error al eliminar la comunidad', error });
   }
